[Vulkan] New Image and ImageView for Swapchain
- Minor refactoring ( using Ref rather than Handle )
- Rewritten Image and ImageView
- Added Framebuffer (incomplete yet)
- Swapchain uses new Image/ImageView API

+++ dali/graphics/vulkan/vulkan-command-buffer.cpp
@@ -298,7 +298,7 @@ struct CommandBuffer::Impl
 
   void DrawIndexed( uint32_t indexCount, uint32_t instanceCount, uint32_t firstIndex, uint32_t vertexOffset, uint32_t firstInstance )
   {
-    mCommandBuffer.drawIndexed( indexCount, instanceCount, firstIndex, vertexOffset, firstInstance );
+    mCommandBuffer.drawIndexed( indexCount, instanceCount, firstIndex, static_cast<int32_t>(vertexOffset), firstInstance );
   }
 
   const std::vector< Handle<VkManaged>> GetResources() const
